square_in_a_matrix.cpp
Given a mxn matrix, count the number of squares in the matrix.squaresinREct

Example 1:

Input:
m = 2, n = 2
Output:
5
Explanation:
There are a total of 5 squares in a 2*2
matrix. Four 1*1 squares and one 2*2 square.
Example 2:

Input:
m = 4, n = 3
Output:
20
Explanation:
There are a total of 20 squares in a 4*3
matrix. It's explained clearly in the
diagram given in the problem statement.
 

Your Task:
You don't need to read input or print anything. Your task is to complete the function squaresInMatrix() which takes 2 Integers m and n as input and returns the number of total squares in a m*n matrix.

 

Expected Time Complexity: O(1)
Expected Auxiliary Space: O(1)

 

Constraints:
1 <= m,n <= 104